Вынес повторяющиеся функции и константы в отдельные модули
