+++ src/const.js
@@ -1,0 +1,7 @@
+export const COLORS = [
+  `black`,
+  `yellow`,
+  `blue`,
+  `green`,
+  `pink`,
+];
