+++ src/mock/task.js
@@ -1,11 +1,5 @@
-// Функция из интернета по генерации случайного числа из диапазона
-// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
+import {getRandomInteger} from "../utils.js";
+import {COLORS} from "../const.js";
 
 const generateDescription = () => {
   const descriptions = [
@@ -48,15 +42,8 @@ const generateRepeating = () => {
 };
 
 const generateColor = () => {
-  const colors = [
-    `black`,
-    `yellow`,
-    `blue`,
-    `green`,
-    `pink`,
-  ];
-  const randomIndex = getRandomInteger(0, colors.length - 1);
-  return (colors[randomIndex]);
+  const randomIndex = getRandomInteger(0, COLORS.length - 1);
+  return (COLORS[randomIndex]);
 };
