+++ src/view/task-edit.js
@@ -1,12 +1,6 @@
-const isExpired = (dueDate) => {
-  if (dueDate === null) {
-    return false;
-  }
-  let currentDate = new Date();
-  currentDate.setHours(23, 59, 59, 999);
+import {COLORS} from "../const.js";
+import {isTaskExpired, isTaskRepeating, humanizeTaskDueDate} from "../utils.js";
 
-  return currentDate.getTime() > dueDate.getTime();
-};
 
 const createTaskEditDateTemplate = (dueDate) => {
   return (
@@ -22,7 +16,7 @@ const createTaskEditDateTemplate = (dueDate) => {
               type="text"
               placeholder=""
               name="date"
-              value="${dueDate.toLocaleString(`en-US`, {day: `numeric`, month: `long`})}"
+              value="${humanizeTaskDueDate(dueDate)}"
             />
           </label>
         </fieldset>`
@@ -31,17 +25,14 @@ const createTaskEditDateTemplate = (dueDate) => {
   );
 };
 
-const isRepeating = (repeating) => {
-  return Object.values(repeating).some(Boolean);
-};
 
 const createTaskEditRepeatingTemplate = (repeating) => {
   return (
     `<button class="card__repeat-toggle" type="button">
-      repeat:<span class="card__repeat-status">${isRepeating(repeating) ? `yes` : `no`}</span>
+      repeat:<span class="card__repeat-status">${isTaskRepeating(repeating) ? `yes` : `no`}</span>
     </button>
 
-    ${isRepeating(repeating)
+    ${isTaskRepeating(repeating)
       ? `<fieldset class="card__repeat-days">
           <div class="card__repeat-days-inner">
             ${Object.entries(repeating).map(([day, repeat]) => `
@@ -63,10 +54,8 @@ const createTaskEditRepeatingTemplate = (repeating) => {
 };
 
 const createTastEditColorsTemplate = (currentColor) => {
-  const colors = [`black`, `yellow`, `blue`, `green`, `pink`];
-
   return (
-    colors.map((color) => {
+    COLORS.map((color) => {
       return (
         `<input
           type="radio"
@@ -102,11 +91,11 @@ export const createTaskEditTemplate = (task = {}) => {
     }
   } = task;
 
-  const deadlineClassName = isExpired(dueDate)
+  const deadlineClassName = isTaskExpired(dueDate)
     ? `card--deadline`
     : ``;
 
-  const repeatingClassName = isRepeating(repeating)
+  const repeatingClassName = isTaskRepeating(repeating)
     ? `card--repeat`
     : ``;
